fix(tasks): VSCode Tasks executed by a compound task do not update state when running #67
